[iOS] Init Microsoft Quick Auth SDK
This patch initializes the Microsoft Quick Auth SDK for iOS, this SDK
is mainly built on top of MSAL for iOS, which enables to leverage the
Microsoft Identity platform. The SDK is a thin layer that wraps the
MSAL, meanwhile, offers a "Sign in with Microsoft" button that can be
customized, e.g., style, theme and text.

The SDK itself can get users into third-party apps quickly and securely,
using their Microsoft Account (MSA), like the Google sign-in does.

Also, this patch includes a sample app under SampleApp/ folder, and
developers can try with it to experience the sign-in button or provided
APIs.

The overall struct of the project looks like:

  ├── MSQASignIn
  │   ├── public
  │   └── src
  └── SampleApp

all the exported headers locate in public/.